Drag and drop
I added a variable to target the small container area and make it the box for dragged items. For now only one song can be dragged at the time as the arrow expression function was made for replacing one song only. I'll try to modify it to make possible to add more songs and queue them. There's still small issues to be fixed.